Added AngularJS
Made infinite scroll controller and factory

Refactored HomeController@next page to work with the new scroll service

+++ resources/views/opus/partials/_opusDetails.blade.php
@@ -51,7 +51,7 @@
                 <div class="panel panel-default details-panel">
                     <div class="panel-heading">
                         Details
-                        @if(Magnus::isOwnerOrHasRole($opus, config('roles.mod-code')))
+                        @if(Auth::check() and Magnus::isOwnerOrHasRole($opus, config('roles.mod-code')))
                             <div class="pull-right operations">
                                 {!! Form::model($opus, ['method'=>'delete', 'class'=>'delete-confirm operations',
                                                        'action'=>['OpusController@destroy', $opus->slug]]) !!}
